fix(android): resolve Kotlin compilation errors in MainActivity, PipeService, and SensorDriver, configure local.properties, and align AGP v8.5.2 compile configurations

+++ android-app/app/src/main/java/org/kyberpipe/client/service/PipeService.kt
@@ -35,7 +35,7 @@ class PipeService : Service() {
             minPollIntervalMs = 500L
         ) { lux, timestamp ->
             try {
-                val sensorJson = uniffi.kyberpipe.createSensorPacket(lux, timestamp.toULong())
+                val sensorJson = uniffi.kyberpipe.createSensorPacket(lux.toDouble(), timestamp.toULong())
                 Log.d("KyberpipeService", "Sensor packet emitted: $sensorJson")
             } catch (e: Exception) {
                 Log.e("KyberpipeService", "Failed to create sensor packet: ${e.message}")
